Revert model to px only
